# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Scootware Forum API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/config:
    get:
      operationId: adminGetConfig
      tags: [admin]
      summary: Get site configuration
      responses:
        "200":
          description: Site config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteConfig"
    patch:
      operationId: adminUpdateConfig
      tags: [admin]
      summary: Update site configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteConfig"
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteConfig"
components:
  schemas:
...
    SiteConfig:
      type: object
      properties:
        siteName:
          type: string
        siteDescription:
          type: string
        maintenanceMode:
          type: boolean
        registrationMode:
          type: string
          enum: [open, invite-only, closed]
        requireEmailVerification:
          type: boolean
        inviteRequestMode:
          type: string
          enum: [admin, auto]
        inviteRequestCooldownDays:
          type: integer
          minimum: 0
        products:
          type: object
          description: >-
            Optional mapping of productId to pricing and bulk settings. Each property
            should be an object with `price`, optional `bulkQuantity`, and optional
            `bulkDiscountPercent` (0-100).
          additionalProperties:
            type: object
            properties:
              price:
                type: number
              bulkQuantity:
                type: integer
              bulkDiscountPercent:
                type: number
            required: [price]
      required: [siteName, siteDescription, maintenanceMode, registrationMode, requireEmailVerification]
...